+++ iocTops/VQM_ITMS/VQM_ITMS_App/src/Exception.h
@@ -44,8 +44,8 @@ std::string epicsShareFunc wcstombs(std::wstring const& WideCharString);
 std::wstring epicsShareFunc mbstowcs(std::string const& MultiByteString);
 
 std::string ToString(double value);
-void FromString(std::string const& String, int& value);
-void FromHexString(std::string const& String, int& value);
+void FromString(std::string const& String, size_t& value);
+void FromHexString(std::string const& String, size_t& value);
 void FromString(std::string const& String, double& value);
 
 #endif //EXCEPTION_H_INCLUDED
